Extract Type typeclass
