+++ src/main/scala/scalaz/distributed/Path.scala
@@ -3,17 +3,17 @@ package scalaz.distributed
 sealed trait Path[A, B]
 
 object Path {
-  final case class Key[K: SupportedType, V: SupportedType](v: K) extends Path[Map[K, V], V]
+  final case class Key[K: Type, V: Type](v: K) extends Path[Map[K, V], V]
 
-  final case class Elements[A: SupportedType]() extends Path[Set[A], A]
+  final case class Elements[A: Type]() extends Path[Set[A], A]
 
-  final case class Composed[A: SupportedType, B: SupportedType, C: SupportedType](
+  final case class Composed[A: Type, B: Type, C: Type](
     x: Path[A, B],
     y: Path[B, C]
   ) extends Path[A, C]
 
-  implicit class PathSyntax[A: SupportedType, B: SupportedType](self: Path[A, B]) {
-    final def >>> [C: SupportedType](that: Path[B, C]): Path[A, C] =
+  implicit class PathSyntax[A: Type, B: Type](self: Path[A, B]) {
+    final def >>> [C: Type](that: Path[B, C]): Path[A, C] =
       Composed(self, that)
   }
 }
