+++ src/main/scala/scalaz/distributed/Protocol.scala
@@ -1,0 +1,5 @@
+package scalaz.distributed
+
+trait Protocol {
+  def access[A: Type](id: MetadataID): Distributed[Metadata]
+}
