+++ src/main/scala/scalaz/distributed/Reified.scala
@@ -1,0 +1,7 @@
+package scalaz.distributed
+
+sealed trait Reified
+
+object Reified {
+  // TODO: type instances
+}
