+++ src/main/scala/scalaz/distributed/Type.scala
@@ -1,0 +1,9 @@
+package scalaz.distributed
+
+trait Type[A] {
+  def reified: Reified
+}
+
+object Type {
+  def apply[A](implicit instance: Type[A]): Type[A] = instance
+}
